luci integration and webui prototype
the webui is experimental, nim's prologue crashes with concurrent ajax
requests. the whole backend needs to be revamped, I suspect the
threadpool is messing with the server file descriptors.

the whole javascript ui is quite primitive too: I should explore the
javascript backend for nim, or move to a more modern but still
lightweight framework.